refactor(hew-lsp): extract text sync handlers
