+++ hew-lsp/src/server/handlers/mod.rs
@@ -1,0 +1,1 @@
+pub(super) mod text_sync;
